Fixed bug in redund where sqlite db was not created when preparing ref set
redund.c
- Moved string allocation to before single outfile branching point. Now
that the file name variables are dynamically allocated, sprintf does not
work to set them so both branches expect initialized buffers to write to.
- Fixed another bug where an error would occur when reading the DEBUG
env var if it is not set.
- Free filename vars

vntrseek.pl
- Removed unused chdirs
- Moved find command invocation directly inside system call

vutil.pm
- Surround call to run_redund with a try block. Drops the database table
so that this load_refprofiles_db can be run again in case of an error.
- In run_redund, add some code to help figure out any error with
retrieving SQL code for table creation. (Might be removed)

+++ perl/lib/vutil.pm
@@ -1,5 +1,5 @@
 package vutil;
-use strict;
+use v5.24;
 use Cwd;
 use DBI;
 use Carp;
@@ -1027,7 +1027,21 @@ sub load_refprofiles_db {
         $dbh->commit;
 
         # Run redund
-        run_redund( $dbh, $refleb36, "reference.leb36", 0, 0 );
+        try {
+            warn "Running redund on ref set\n";
+            run_redund( $dbh, $refleb36, "reference.leb36", 0, 0 );
+        }
+        catch {
+            warn "Error running redund on reference set: $_\n";
+            # if (/DBD::SQLite::db/) {
+            # }
+
+            $dbh->rollback;
+            $dbh->do(q{DROP TABLE IF EXISTS `ref_profiles`});
+            $dbh->disconnect;
+
+            die "\n";
+        };
         return 1;
     }
 
@@ -1178,7 +1192,15 @@ sub run_redund {
     my ($minrep_sql) = $dbh->selectrow_array(
         q{SELECT sql FROM minrep.sqlite_master
         WHERE name = 'minreporder'}
-    ) or carp "Couldn't do statement: $DBI::errstr\n";
+    );
+    
+    unless ($minrep_sql) {
+        my $res = $dbh->selectall_arrayref(q{SELECT * FROM minrep.sqlite_master});
+        use Data::Dumper;
+        say Dumper($res);
+        die "unable to get CREATE TABLE statement for minreporder from redund db\n";
+    }
+
     $minrep_sql =~ s/minreporder/main.minreporder/;
     $dbh->do(q{DROP TABLE IF EXISTS main.minreporder})
         or carp "Couldn't do statement: $DBI::errstr\n";
